+++ tests/fields/test_time.py
@@ -62,13 +62,13 @@ class TestDatetimeFields(test.TestCase):
         now = timezone.now()
         obj0 = await testmodels.DatetimeFields.create(datetime=now)
         values = await testmodels.DatetimeFields.get(id=obj0.id).values("datetime")
-        self.assertEqual(values[0]["datetime"], now)
+        self.assertEqual(values["datetime"], now)
 
     async def test_values_list(self):
         now = timezone.now()
         obj0 = await testmodels.DatetimeFields.create(datetime=now)
         values = await testmodels.DatetimeFields.get(id=obj0.id).values_list("datetime", flat=True)
-        self.assertEqual(values[0], now)
+        self.assertEqual(values, now)
 
     async def test_get_utcnow(self):
         now = datetime.utcnow().replace(tzinfo=get_default_timezone())
@@ -171,13 +171,13 @@ class TestDateFields(test.TestCase):
         today = date.today()
         obj0 = await testmodels.DateFields.create(date=today)
         values = await testmodels.DateFields.get(id=obj0.id).values("date")
-        self.assertEqual(values[0]["date"], today)
+        self.assertEqual(values["date"], today)
 
     async def test_values_list(self):
         today = date.today()
         obj0 = await testmodels.DateFields.create(date=today)
         values = await testmodels.DateFields.get(id=obj0.id).values_list("date", flat=True)
-        self.assertEqual(values[0], today)
+        self.assertEqual(values, today)
 
     async def test_get(self):
         today = date.today()
@@ -217,7 +217,7 @@ class TestTimeDeltaFields(test.TestCase):
             timedelta=timedelta(days=35, seconds=8, microseconds=1)
         )
         values = await testmodels.TimeDeltaFields.get(id=obj0.id).values("timedelta")
-        self.assertEqual(values[0]["timedelta"], timedelta(days=35, seconds=8, microseconds=1))
+        self.assertEqual(values["timedelta"], timedelta(days=35, seconds=8, microseconds=1))
 
     async def test_values_list(self):
         obj0 = await testmodels.TimeDeltaFields.create(
@@ -226,7 +226,7 @@ class TestTimeDeltaFields(test.TestCase):
         values = await testmodels.TimeDeltaFields.get(id=obj0.id).values_list(
             "timedelta", flat=True
         )
-        self.assertEqual(values[0], timedelta(days=35, seconds=8, microseconds=1))
+        self.assertEqual(values, timedelta(days=35, seconds=8, microseconds=1))
 
     async def test_get(self):
         delta = timedelta(days=35, seconds=8, microseconds=1)