+++ CHANGELOG.rst
@@ -4,15 +4,25 @@
 Changelog
 =========
 
-
 .. rst-class:: emphasize-children
+
+0.18
+====
+0.18.0
+------
+Added
+^^^^^
+- Add Case-When support. (#943)
+- Add `Rand`/`Random` function in contrib. (#944)
+Changed
+^^^^^^^
+- Move `Function`, `Aggregate` from `functions.py` to `expressions.py`. (#943)
+- Move `Q` from `query_utils.py` to `expressions.py`.
 
 0.17
 ====
-
 0.17.8
 ------
-
 Added
 ^^^^^
 - Add `Model.raw` method to support the raw sql query.
@@ -24,9 +34,11 @@ Fixed
 - Fix `QuerySet` subclass being lost when `_clone` is run on the instance.
 - Fix bug in `.values` with `source_field`. (#844)
 - Fix `contrib.blacksheep` exception handlers, use builtin json response. (#914)
+- Fix Indexes defined in Meta class do not make use of `exists` parameter in their template (#928)
 Changed
 ^^^^^^^
 - Allow negative values with `IntEnumField`. (#889)
+- Make `.values()` and `.values_list()` awaited return more consistent. (#899)
 
 0.17.7
 ------